% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wfs.R
\name{describe_protocols}
\alias{describe_protocols}
\title{Summarise the link protocols present in a catalogue}
\usage{
describe_protocols(x, max_shown = 10L)
}
\arguments{
\item{x}{character vector of \code{link_protocol} values, \code{NA}s allowed.}

\item{max_shown}{number of protocols to name before summarising the rest.}
}
\value{
character of length one, e.g. \verb{"INSPIRE ATOM (826), OGC:WMS (776)"}.
}
\description{
Summarise the link protocols present in a catalogue
}
\keyword{internal}
