% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_wfs.R
\name{empty_wfs_result}
\alias{empty_wfs_result}
\title{Empty result of find_wfs()}
\usage{
empty_wfs_result()
}
\value{
zero-row tibble with the columns of \code{\link[=find_wfs]{find_wfs()}}.
}
\description{
Empty result of find_wfs()
}
\keyword{internal}
