% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing helper}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Any object that might be \code{NULL} or of length 0.}

\item{y}{Fallback value to be returned when \code{x} is \code{NULL} or length 0.}
}
\value{
Either \code{x} (when present) or \code{y} (when \code{x} is missing).
}
\description{
Small utility to replace \code{NULL} or zero-length objects by a default value.
This is handy when parsing GeoNetwork JSON/XML responses where some
elements are simply missing.
}
\keyword{internal}
